[PATCH] 2.5.54 AGP driver fixes for x86-64
This is just to make x86-64 boot with IOMMU and compiled in AGP driver
again.  Previously both AGP and the IOMMU code would set up an aperture,
with AGP being winner, messing up pci_map_*.

This is needed for the x86-64 update for 2.5.54 I sent earlier.

Actually using /dev/agpgart still corrupts memory unfortunately, haven't
tracked that down yet.

I also commented on some (usually 64bit, but at least one 32bit) bugs in
the driver too.

These changes should be all nops on i386.

 - Export amd k8 init function for x86-64.  It needs to be called early
   by the iommu code.  Add some comments to warn about about that (the
   upcoming VIA and SIS K8 chipsets will have the same problem)
 - Declare gatt table as u32 following the AGP spec.
 - Comment some bugs.

+++ drivers/char/agp/agp.h
@@ -128,8 +128,8 @@ struct agp_bridge_data {
 	void *dev_private_data;
 	struct pci_dev *dev;
 	struct gatt_mask *masks;
-	unsigned long *gatt_table;
-	unsigned long *gatt_table_real;
+	u32 *gatt_table;
+	u32 *gatt_table_real;
 	unsigned long scratch_page;
 	unsigned long gart_bus_addr;
 	unsigned long gatt_bus_addr;
